Added again asynchronous calls.
Add again asynchronous calls in DDS.

+++ rpcddsgen/src/com/eprosima/rpcdds/RPCDDSGEN.java
@@ -71,6 +71,8 @@ public class RPCDDSGEN
     private String m_languageOption = "C++";
     private boolean m_ppDisable = false;
     private boolean m_replace = false;
+    //! Add information to use fastrpcgen internally. 
+    private boolean m_local = false;
     private String m_ppPath = null;
     private String m_protocol = "dds"; // Default protocol -> DDS
     
@@ -224,6 +226,10 @@ public class RPCDDSGEN
                 printHelp();
                 System.exit(0);
             }
+            else if(arg.equals("-local"))
+            {
+                m_local = true;
+            }
             // Get include directories
             else if(arg.equals("-I"))
             {
@@ -348,6 +354,9 @@ public class RPCDDSGEN
 	        	project.addCommonSrcFile("MessageHeaderSupport.cxx");
 	        	project.setUnique(true);
 	        	solution.addProject(project);
+
+                // TODO Review
+                solution.addLibrary("boost_system");
             }
             catch(Exception ex)
             {
@@ -357,14 +366,14 @@ public class RPCDDSGEN
         }
         else if(m_protocol.equalsIgnoreCase("rest"))
         {
-            solution.addLibrary("boost_system-mt");
-            solution.addLibrary("boost_thread-mt");
+            solution.addLibrary("boost_system");
+            solution.addLibrary("boost_thread");
         }
         else if(m_protocol.equalsIgnoreCase("fastcdr"))
         {
             solution.addInclude("$(RPCDDSHOME)/include/protocols/cdr");
-            solution.addLibrary("boost_system-mt");
-            solution.addLibrary("boost_thread-mt");
+            solution.addLibrary("boost_system");
+            solution.addLibrary("boost_thread");
             solution.addLibrary("cdr");
         }
         
@@ -541,7 +550,7 @@ public class RPCDDSGEN
         if(idlParseFileName != null)
         {
 	        // Create initial context.
-	        Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode);
+	        Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode, "rest");
 	        
 	        // Create template manager
 	        TemplateManager tmanager = new TemplateManager("com/eprosima/rpcdds/idl/templates");
@@ -716,7 +725,7 @@ public class RPCDDSGEN
         if(idlParseFileName != null)
         {
             // Create initial context.
-            Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode);
+            Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode, "cdr");
             
             // Create template manager
             TemplateManager tmanager = new TemplateManager("com/eprosima/rpcdds/idl/templates");
@@ -902,7 +911,7 @@ public class RPCDDSGEN
         if(idlParseFileName != null)
         {
 	        // Create initial context.
-	        Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode);
+	        Context ctx = new Context(onlyFileName, idlFilename, m_includePaths, m_clientcode, m_servercode, "dds");
 	        
 	        // Create template manager
 	        TemplateManager tmanager = new TemplateManager("com/eprosima/rpcdds/idl/templates");
@@ -924,8 +933,8 @@ public class RPCDDSGEN
 	        // Load template to generate example to use Proxies.
 	        tmanager.addGroup("DDSClientExample");
 	        // Load template to generate proxy async support files.
-	        //tmanager.addGroup("AsyncSupportHeader");
-	        //tmanager.addGroup("AsyncSupportSource");
+	        tmanager.addGroup("DDSAsyncSupportHeader");
+	        tmanager.addGroup("DDSAsyncSupportSource");
 	        // Load template to generate Server for topics.
 	        tmanager.addGroup("ServerHeader");
 	        tmanager.addGroup("ServerSource");
@@ -1016,16 +1025,24 @@ public class RPCDDSGEN
 		            
 		            if(returnedValue = Utils.writeFile(m_tempDir + onlyFileName + "RequestReply.idl", maintemplates.getTemplate("TopicsIDL"), true))
 		            {	
-		                if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "Protocol.h", maintemplates.getTemplate("ProtocolHeader"), m_replace))
+                        if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSAsyncSupport.h", maintemplates.getTemplate("DDSAsyncSupportHeader"), m_replace))
                         {
-    		                if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSProtocol.h", maintemplates.getTemplate("DDSProtocolHeader"), m_replace))
-    		                {
-    		                    returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSProtocol.cxx", maintemplates.getTemplate("DDSProtocolSource"), m_replace);
-    		                    
-    		                    project.addCommonIncludeFile(onlyFileName + "Protocol.h");
-    		                    project.addCommonIncludeFile(onlyFileName + "DDSProtocol.h");
-    		                    project.addCommonSrcFile(onlyFileName + "DDSProtocol.cxx");
-    		                }
+                            if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSAsyncSupport.cxx", maintemplates.getTemplate("DDSAsyncSupportSource"), m_replace))
+                            {
+                                if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "Protocol.h", maintemplates.getTemplate("ProtocolHeader"), m_replace))
+                                {
+                                    if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSProtocol.h", maintemplates.getTemplate("DDSProtocolHeader"), m_replace))
+                                    {
+                                        returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "DDSProtocol.cxx", maintemplates.getTemplate("DDSProtocolSource"), m_replace);
+
+                                        project.addCommonIncludeFile(onlyFileName + "Protocol.h");
+                                        project.addCommonIncludeFile(onlyFileName + "DDSProtocol.h");
+                                        project.addCommonSrcFile(onlyFileName + "DDSProtocol.cxx");
+                                        project.addCommonIncludeFile(onlyFileName + "DDSAsyncSupport.h");
+                                        project.addCommonSrcFile(onlyFileName + "DDSAsyncSupport.cxx");
+                                    }
+                                }
+                            }
                         }
 		            }
 		        }
@@ -1037,19 +1054,11 @@ public class RPCDDSGEN
 		            {
 		                if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "Proxy.cxx", maintemplates.getTemplate("ProxySource"), m_replace))
 		                {
-                            //if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "AsyncSupport.h", maintemplates.getTemplate("AsyncSupportHeader"), m_replace))
-                            {
-                                //if(returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "AsyncSupport.cxx", maintemplates.getTemplate("AsyncSupportSource"), m_replace))
-                                {
-                                	project.addClientIncludeFile(onlyFileName + "Proxy.h");
-                                	project.addClientSrcFile(onlyFileName + "Proxy.cxx");;
-                                	//project.addClientIncludeFile(onlyFileName + "AsyncSupport.h");
-                                	//project.addClientSrcFile(onlyFileName + "AsyncSupport.cxx");
-                                	
-                                    if(m_exampleOption != null)
-                                        returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "ClientExample.cxx", maintemplates.getTemplate("DDSClientExample"), m_replace);
-                                }
-                            }
+                            project.addClientIncludeFile(onlyFileName + "Proxy.h");
+                            project.addClientSrcFile(onlyFileName + "Proxy.cxx");;
+
+                            if(m_exampleOption != null)
+                                returnedValue = Utils.writeFile(m_outputDir + onlyFileName + "ClientExample.cxx", maintemplates.getTemplate("DDSClientExample"), m_replace);
 		                }
 		            }
 		        }
@@ -1638,6 +1647,7 @@ public class RPCDDSGEN
             makecxx.setAttribute("example", m_exampleOption);
     		makecxx.setAttribute("arch", arch);
     		makecxx.setAttribute("version", getVersion());
+            makecxx.setAttribute("local", m_local);
             
             returnedValue = Utils.writeFile(m_outputDir + "makefile_" + m_exampleOption, makecxx, m_replace);
     	}
